[VIEW IMPROVEMENT] CSS added for listing recipes in grid on main page floating left.
Added some css into the style.scss so that the recipes on the main page are displayed as 3 column grids. They float to the left so that they read new to old, left to right. Also aligned the text to be central but this may need changing.
Updated the defaultcontent to be more markety. This is just placeholder code that is being used for testing though (simulating multiple recipes by pushing more on to the array).