[IMPROVEMENT] Improved the build system by adding noserver option. [DOCUMENTATION] Improved documentation by explaining build system.
Previously, the default option of the build system would compile the scss,
run jshint to check the javascript and then launch the server.

This is annoying when developing just the frontend scss stuff and you want
a fast feedback loop of dev->test->dev->test... to see what changes look
like.

Now, the build option of noserver can be specified and built as "grunt noserver"
which does the same as default but without launching the server.
This mean you can start the server in one terminal then work on the frontend
stuff from another.