[UI IMPROVEMENT] Added UI feature where user can add ingrediants and quantities when creating recipes.
Previously, the create page had little to no functionality. You could add
a text box by clicking the button and that was it.

Now, it is barely functional but the basics of the UI are there.
Clicking add an ingrediant adds a text box for the ingrediant name and a
text box for the quantity. There's also a submit button but it currently
doesn't do anything.